Implement adding a new key on other device
The new function add_new_key_other_device() calls
/api/rpc/init_credential just like init_credential() still does,
but instead of invoking navigator.credentials.create(),
it just displays a URL with the challenge to the user,
allowing that URL to be loaded from a different device,
to continue the credential creation process,
and then finally call store_credential() to store it
to the server.

Also remove FUNCTIONS/hash_extension.sql as it has been moved
to a new separate project.