Implement application level RBAC and auto-create UI from Swagger OpenAPI
RBAC stands for Role-Based Access Control and is a standard
way to handle Authorization.

An automatic User Interface is now derived from the
Swagger OpenAPI JSON provided by PostgREST,
which means when developing locally,
when a VIEW or FUNCTION is CREATEed in the "api"
schema, they will appear automatically in the front-end,
without any need to populate any lookup-tables
or write any specific Javascript/HTML for each resource.

The default look-and-feel for the automatic user-inteface
is not intended to be used as-is in production,
but speeds up the development cycles, as backend
developers can then immediately play with their
newly created resources and get a feeling for if
the work is sufficiently ready to be handed over
to front-end developers for a proper integration
with custom Javascript and/or HTML and/or CSS.

Implementation details:

The following line has been added to postgrest.conf:

  pre-request = "public.auth"

This causes the new function public.auth() to be invoked
before any resource is accessed, such as a view or function.

The new auth() function now is responsible for both
Authentication and Authorization.

Authentication works almost like before,
except the work to verify the access_token cookie in the
"request.cookie.access_token" GUC has now moved
into the auth() function instead of being handled
directly in the user_id() function, which is instead
rely on the new GUC "uniphant.user_id", set by auth()
to the authenticated user_id.

Authorization works by checking if the HTTP/HTTPS
request path passed by PostgREST via the
GUC "request.path" matches a resource's resource_path
in the new resources table,
and if the current user has access to it
via any of the roles it has been granted.

api.openapi_swagger()

  Returns the OpenAPI Swagger JSON API spec
  from PostgREST as-is, BUT also filters out
  any resources (paths) the current user
  doesn't have access to.

check_resource_access(_resource_id integer)

  Helper-function used by:
    auth() to check if the user is authorized to access the resource
    openapi_swagger() to filter out any denied resources
    user_resources view, to show a list of available resources

auto_add_new_resources()

  The function automatically register new functions and views
  as resources. By also adding the "anonymous" user
  to the special "*" role, such resources will appear
  automatically in the front-end.
  Meant to be used only when developing locally.

New objects:

+  TABLES/roles.sql
+  TABLES/resources.sql
+  TABLES/permissions.sql
+  TABLES/credentials.sql
+  TABLES/role_memberships.sql
+  TABLES/openapi_swagger.sql
+  VIEWS/api/users.sql
+  VIEWS/api/current_user.sql
+  VIEWS/api/users.sql
+  VIEWS/api/resources.sql
+  VIEWS/api/roles.sql
+  VIEWS/api/user_credentials.sql
+  VIEWS/api/user_resources.sql
+  VIEWS/api/user_role_memberships.sql
+  VIEWS/api/permissions.sql
+  VIEWS/api/credentials.sql
+  VIEWS/api/role_memberships.sql
+  FUNCTIONS/register_resource.sql
+  FUNCTIONS/check_resource_access.sql
+  FUNCTIONS/auth.sql
+  FUNCTIONS/set_user_id.sql
+  FUNCTIONS/has_role.sql
+  FUNCTIONS/set_openapi_swagger.sql
+  FUNCTIONS/api/sign_in.sql
+  FUNCTIONS/api/sign_up.sql
+  FUNCTIONS/api/create_role.sql
+  FUNCTIONS/api/grant_role_to_user.sql
+  FUNCTIONS/api/grant_resource_to_role.sql
+  FUNCTIONS/api/create_user.sql
+  FUNCTIONS/api/update_credential_validity.sql
+  FUNCTIONS/api/openapi_swagger.sql
+  FUNCTIONS/notify_ddl_postgrest.sql
+  FUNCTIONS/auto_add_new_resources.sql