Use PostgreSQL to keep the state of processes
The biggest change in this commit is to entrust PostgreSQL to keep the state of processes,
to prevent multiple instances of the same worker_id from  running simultaneously.

The PID is now registered upon start of a worker in the processes table along with the process_id UUID.

The new get_existing_process_info() database function returns the process_id UUID and PID for a worker_id.
This function is used to check if the alleged process is already running on the host, or if it has died,
allowing clean-up and restart, if desired.

The reason for this move is simplicity. There is a lot of legacy and complication in POSIX file locking and
signal handling, which might not be trivial to implement in the same way in different languages/platforms.
Furthermore, since we're already by-design dependent on the availability of the PostgreSQL database,
we might as well entrust it with this task as well. The atomicity in ACID is exactly what we need.

Also drop the idea of implicitly passing the process_id via application_name, that was a stupid idea since it
won't work for those who use connection pooling.