Simplify process management by requesting process termination via the database.
The process clean-up code in the worker() function that checked if the existing pid
was alive, and if not, cleaned up by deleting the process in the database,
has now been removed.

Instead, a new boolean column has been added, processes.termination_requested,
which allows us to gracefully request termination of a running worker process,
and wait for itself to stop and delete its row in processes to acknowledge termination.

Also use named parameters when calling database functions to improve readability
of database log files, since otherwise it's not obvious what UUID is what,
and it also provides better safety, but perhaps not in Python3 since
variables are passed positional.